* Renamed source files for Neo Editor
