+++ Neo/NeoEditor/Bins/MeshBin.cpp
@@ -1,10 +1,6 @@
-/////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Maratis
-// MMeshBin.cpp
-/////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 //========================================================================
-//  Maratis, Copyright (c) 2003-2012 Anael Seghezzi <www.maratis3d.com>
+//  Copyright (c) 2003-2012 Anael Seghezzi <www.maratis3d.com>
+//	Copyright (c) 2015 Yannick Pflanzer <www.neo-engine.de>
 //
 //  This program is free software; you can redistribute it and/or
 //  modify it under the terms of the GNU General Public License
@@ -23,7 +19,7 @@
 //========================================================================
 
 #include <NeoEngine.h>
-#include "MMeshBin.h"
+#include "MeshBin.h"
 
 using namespace Neo;
