+++ Neo/NeoEditor/FLUID/Callbacks.cpp
@@ -1,3 +1,22 @@
+//========================================================================
+//	Copyright (c) 2014-2015 Yannick Pflanzer <www.neo-engine.de>
+//
+//  This program is free software; you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License
+//  as published by the Free Software Foundation; either version 2
+//  of the License, or (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software Foundation,
+//  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+//
+//========================================================================
+
 #include <HTTPRequest.h>
 
 #include "Callbacks.h"
@@ -30,8 +49,8 @@
 
 using namespace Neo;
 
-#include "../MFilesUpdate/MFilesUpdate.h"
-#include "../MLoaders/MAssimpMeshLoader.h"
+#include "../FilesUpdate/FilesUpdate.h"
+#include "../Loaders/AssimpMeshLoader.h"
 #include <MCore.h>
 #include "MainWindow.h"
 #include "Translator.h"
