+++ Neo/NeoEditor/FLUID/GLBox.cxx
@@ -1,3 +1,22 @@
+//========================================================================
+//	Copyright (c) 2014-2015 Yannick Pflanzer <www.neo-engine.de>
+//
+//  This program is free software; you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License
+//  as published by the Free Software Foundation; either version 2
+//  of the License, or (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software Foundation,
+//  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+//
+//========================================================================
+
 #include "GLBox.h"
 #include <MWindow.h>
 #include <NeoEngine.h>
@@ -5,11 +24,10 @@
 #include <MKeyboard.h>
 
 #include "../MWindow/MMouse.h"
-#include "../MFilesUpdate/MFilesUpdate.h"
 #include "../Maratis/Maratis.h"
 #include "Callbacks.h"
 #include <MCore.h>
-#include "../MRenderArray/MRenderArray.h"
+#include "../RenderArray/RenderArray.h"
 #include "MainWindow.h"
 
 // NeoGui
@@ -19,12 +37,13 @@
 #include <FL/Fl_Int_Input.H>
 #include <FL/Fl_Double_Window.H>
 
-#include "../MPluginScript/MPluginScript.h"
+#include "../PluginScript/PluginScript.h"
 
 #include <algorithm>
 #include <string>
 
 #include <Shiny.h>
+#include "../FilesUpdate/FilesUpdate.h"
 
 using namespace Neo;
