+++ Neo/NeoEditor/FilesUpdate/FilesUpdate.cpp
@@ -26,8 +26,8 @@
 #include <time.h>
 #include <sys/types.h>
 #include <sys/stat.h>
+#include "FilesUpdate.h"
 
-#include "MFilesUpdate.h"
 
 using namespace Neo;
